etls/twitter/requirements.txt
attrs==21.2.0
backoff==1.11.1
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
cymem==2.0.5
de-core-news-sm==3.1.0
en-core-web-sm==3.1.0
es-core-news-sm==3.1.0
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
flask-swagger==0.2.14
fr-core-news-sm==3.1.0
gunicorn==20.1.0
huggingface-hub==0.0.12
hunspell==0.5.5
idna==3.2
it-core-news-sm==3.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.2
orion==1.0.0
packaging==21.0
pandas==1.3.2
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
pt-core-news-sm==3.1.0
pydantic==1.8.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
ratelimit==2.2.1
regex==2021.8.21
requests==2.26.0
sacremoses==0.0.45
sentiment==1.0.0
six==1.16.0
smart-open==5.2.0
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
textdistance==4.2.1
thinc==8.0.8
tokenizers==0.10.3
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tqdm==4.62.1
transformers==4.9.2
tweet-preprocessor==0.6.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
